requirements.txt
telethon
python-dotenv
psycopg2-binary
dbt-postgres
fastapi
uvicorn
sqlalchemy
pandas
dagster
dagit
ultralytics
torch
